backend/requirements.txt
asgiref==3.6.0
boto3==1.26.138
botocore==1.29.138
coverage==7.2.5
Django==4.1
django-cors-headers==3.14.0
django-s3-storage==0.14.0
djangorestframework==3.14.0
jmespath==1.0.1
psycopg2==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15